Java基本服务运行起来了，但是存在和c#一样的超时错误。BUG。

+++ ZezexJava/server/src/Game/Login/Onlines.java
@@ -472,7 +472,7 @@ public class Onlines {
 			}
 		}
 
-		public final int ProcessLinkConfirm(String linkName) {
+		public final long ProcessLinkConfirm(String linkName) {
 			synchronized (this) {
 				getLinkNames().remove(linkName);
 				if (getLinkNames().isEmpty()) {
